refactor: Moved customInput to common widget directory #5
