+++ lib/widgets/membersAdminEdit/membersEditForm.dart
@@ -1,3 +1,4 @@
+import 'package:calendar_app/common/customInput.dart';
 import 'package:calendar_app/constants/customColor.dart';
 import 'package:calendar_app/utils/toastMessage.dart';
 import 'package:flutter/material.dart';
@@ -60,47 +61,6 @@ class _FormComponentState extends State<MembersEditForm> {
             ),
             child: const Text('Submit'),
           ),
-        ],
-      ),
-    );
-  }
-}
-
-class CustomInput extends StatelessWidget {
-  const CustomInput({
-    super.key,
-    required this.label,
-    required this.onSaved,
-    required this.validator,
-  });
-
-  final String label;
-  final FormFieldSetter onSaved;
-  final FormFieldValidator validator;
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 300,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            label,
-            style: const TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.bold,
-              color: CustomColor.darkGray,
-            ),
-          ),
-          TextFormField(
-            onSaved: onSaved,
-            validator: validator,
-            // autovalidateMode: AutovalidateMode.always, // 제어컴포넌트
-          ),
-          const SizedBox(
-            height: 16.0,
-          )
         ],
       ),
     );
