+++ lib/widgets/myPageEdit/myPageEditForm.dart
@@ -1,3 +1,5 @@
+import 'package:calendar_app/common/customInput.dart';
+import 'package:calendar_app/common/customSecretInput.dart';
 import 'package:calendar_app/constants/customColor.dart';
 import 'package:calendar_app/utils/toastMessage.dart';
 import 'package:flutter/material.dart';
@@ -59,110 +61,6 @@ class _FormComponentState extends State<MyPageEditForm> {
             ),
             child: const Text('Submit'),
           ),
-        ],
-      ),
-    );
-  }
-}
-
-class CustomSecretInput extends StatefulWidget {
-  const CustomSecretInput({
-    super.key,
-    required this.label,
-    required this.onSaved,
-    required this.validator,
-  });
-
-  final String label;
-  final FormFieldSetter onSaved;
-  final FormFieldValidator validator;
-
-  @override
-  State<CustomSecretInput> createState() => _CustomSecretInputState();
-}
-
-class _CustomSecretInputState extends State<CustomSecretInput> {
-  bool _secret = true;
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 300,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            widget.label,
-            style: const TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.bold,
-              color: CustomColor.darkGray,
-            ),
-          ),
-          TextFormField(
-            obscureText: _secret,
-            onSaved: widget.onSaved,
-            validator: widget.validator,
-            // autovalidateMode: AutovalidateMode.always, // 제어컴포넌트
-            decoration: InputDecoration(
-              suffixIcon: GestureDetector(
-                child: _secret
-                    ? const Icon(Icons.visibility_outlined)
-                    : const Icon(Icons.visibility_off_outlined),
-                onTap: () => setState(() {
-                  _secret = !_secret;
-                }),
-              ),
-            ),
-          ),
-          const SizedBox(
-            height: 16.0,
-          )
-        ],
-      ),
-    );
-  }
-}
-
-class CustomInput extends StatelessWidget {
-  const CustomInput({
-    super.key,
-    required this.label,
-    required this.onSaved,
-    required this.validator,
-  });
-
-  final String label;
-  final FormFieldSetter onSaved;
-  final FormFieldValidator validator;
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 300,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            label,
-            style: const TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.bold,
-              color: CustomColor.darkGray,
-            ),
-          ),
-          TextFormField(
-            onSaved: onSaved,
-            validator: validator,
-            // autovalidateMode: AutovalidateMode.always, // 제어컴포넌트
-          ),
-          const SizedBox(
-            height: 16.0,
-          )
         ],
       ),
     );
